requirements.txt
pygame==2.6.0
pillow==10.4.0